scattered files into different files
