+++ devices.json
@@ -1,0 +1,9 @@
+{
+    "selected": "Alex's Phone",
+    "devices": {
+        "Alex's Phone": {
+            "username": "YBWMYT",
+            "password": "q3psgbpimxnmtt"
+        }
+    }
+}
